+++ frontend/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "bootstrap": "^4.3.1",
     "history": "^4.6.3",
+    "jquery": "^3.4.1",
     "react": "^16.0.0",
     "react-bootstrap": "^1.0.0-beta.14",
     "react-dom": "^16.0.0",
