+++ frontend/src/HomePage/HomePage.jsx
@@ -3,41 +3,22 @@ import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import { userActions } from '../_actions';
+import { userService } from '../_services';
 
 class HomePage extends React.Component {
-    componentDidMount() {
-        this.props.getUsers();
-    }
-
-    handleDeleteUser(id) {
-        return (e) => this.props.deleteUser(id);
-    }
+   // componentDidMount() {   this.props.getProfile() }
 
     render() {
         const { user, users } = this.props;
         return (
             <div className="col-md-6 col-md-offset-3">
-                <h1>Hi {user.firstName}!</h1>
+                <h1>Hi {user.}!</h1>
                 <p>You're logged in with React!!</p>
                 <h3>All registered users:</h3>
                 {users.loading && <em>Loading users...</em>}
                 {users.error && <span className="text-danger">ERROR: {users.error}</span>}
-                {users.items &&
-                    <ul>
-                        {users.items.map((user, index) =>
-                            <li key={user.id}>
-                                {user.firstName + ' ' + user.lastName}
-                                {
-                                    user.deleting ? <em> - Deleting...</em>
-                                    : user.deleteError ? <span className="text-danger"> - ERROR: {user.deleteError}</span>
-                                    : <span> - <a onClick={this.handleDeleteUser(user.id)}>Delete</a></span>
-                                }
-                            </li>
-                        )}
-                    </ul>
-                }
                 <p>
-                    <Link to="/login">Logout</Link>
+                    <Link to="/opening">Logout</Link>
                 </p>
             </div>
         );
@@ -51,8 +32,7 @@ function mapState(state) {
 }
 
 const actionCreators = {
-    getUsers: userActions.getAll,
-    deleteUser: userActions.delete,
+    getProfile: userService.getProfile
 }
 
 const connectedHomePage = connect(mapState, actionCreators)(HomePage);
