+++ frontend/src/OpeningPage/login/index.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { history } from '../../_helpers';
 import { userActions } from '../../_actions';
