+++ frontend/src/_actions/user.actions.js
@@ -20,7 +20,7 @@ function login(username, password) {
                 user => {
                     console.log(user);
                     dispatch(success(user));
-                    // history.push('/');
+                    history.push('/');
                 },
                 error => {
                     console.log(error.toString());
