+++ frontend/src/_services/user.service.js
@@ -33,6 +33,14 @@ function logout() {
     localStorage.removeItem('user');
 }
 
+function getProfile(){
+    const requestOptions = {
+        method: 'GET',
+        headers: authHeader()
+    };
+    return fetch(`${config.apiUrl}/member/profile`, requestOptions).then(handleResponse);
+}
+
 function getAll() {
     const requestOptions = {
         method: 'GET',
@@ -84,12 +92,6 @@ function handleResponse(response) {
     return response.text().then(text => {
         const data = text && JSON.parse(text);
         if (!response.ok) {
-            if (response.status === 401) {
-                // auto logout if 401 response returned from api
-                logout();
-                location.reload(true);
-            }
-
             const error = (data && data.message) || response.statusText;
             return Promise.reject(error);
         }
