+++ frontend/src/index.jsx
@@ -5,10 +5,6 @@ import { Provider } from 'react-redux';
 import { store } from './_helpers';
 import { App } from './App';
 
-// setup fake backend
-import { configureFakeBackend } from './_helpers';
-configureFakeBackend();
-
 render(
     <Provider store={store}>
         <App />
